Binder: Added binds
You may now create binds for commands using the following syntax :

/bind /<bind call> /<command to be binded>

and call the bind the following way

/<bind call>

Example, say we wanted to bind /audio play TeamFortress2 heavy_incoming  to /incoming

/bind /incoming /audio play TeamFortress2 heave_incoming

This will return the following message :

Bot: Successfully binded "audio play TeamFortress2 heavy_incoming" to "incoming"

Now all it takes is to call /incoming and we'll get the heavy_incoming audio file

/incoming

Bot: tf2_heavy_incoming.mp3 |
                            V

IMPORTANT NOTE 1: Bind are stored in RAM and wiped on server shutdown (that is killing TimeForUpdates). This means that all binds are wiped on a restart. We're
planning to support perma binds soon

IMPORTANT NOTE 2: Binds do NOT serve as aliases, hence they will not support additional arguments

The following example is NOT possible

/bind /sound /audio

/sound list

The reason for this is as binding allows users to create binds that use whitespaces across them.
While arguments could be implemented with some programming effort, it is simply not the way we meant to design binds.

Signed-off-by: CTXz <ctx.xda@gmail.com>